nineth commit

+++ c++/function.h
@@ -2,4 +2,4 @@
 using namespace std;
 
 void calcul(int a, int b);
-int factoriel(int n);
+long double factoriel(int n);
